Restructure folder structure.
